データベース周りの処理を修正。 - データ管理: python manage_db.py --init --demo でいつでもリセット＆ダミー投入が可能。公開時は --init だけ実行すればOK - 過去データ: バックエンドAPIの修正により、掲示板や時間割からは自動的に消える
