--- demo_data_generator.py
@@ -1,262 +1,0 @@
-import os
-import sqlite3
-import psycopg2
-import random
-import json
-from datetime import datetime, timedelta
-from werkzeug.security import generate_password_hash
-
-# --- 設定 ---
-# ここに Render の "External Database URL" を貼り付けるか、環境変数から読み込む
-# 例: "postgres://user:pass@host/dbname"
-TARGET_DB_URL = os.environ.get("DATABASE_URL") 
-# TARGET_DB_URL = "ここにコピーしたURLを直接貼っても動きます"
-
-DB_PATH = "coma_link.db" # ローカルの場合のファイル名
-RANDOM_SEED = 12345
-NUM_USERS = 50
-NUM_RECRUITMENTS = 300
-# ------------
-
-random.seed(RANDOM_SEED)
-
-# 固定データセット (変更なし)
-FACULTIES = ['工学部', '情報理工学部', '文学部', '法学部', '経済学部', '理学部', '農学部', '芸術学部']
-CIRCLES = ['テニス', 'サッカー', '軽音', 'プログラミング', '美術', '吹奏楽', 'ダンス', 'ボランティア', '茶道']
-CATEGORIES = ['食事', '勉強', 'スポーツ', '雑談', '趣味', 'その他']
-LOCATIONS = ['A-101', '図書館', '学食', 'カフェテリア', '体育館', '部室棟', '芝生広場', '駅前']
-TITLES_BY_CAT = {
-    '食事': ['ランチ行きませんか', '学食でご飯', 'ラーメン食べたい', 'カフェで休憩'],
-    '勉強': ['課題手伝って', 'テスト勉強', '図書館で自習', 'プログラミング教えます'],
-    'スポーツ': ['バスケしよう', 'キャッチボール', 'テニス相手募集', 'ランニング'],
-    '雑談': ['暇つぶし', '雑談しましょう', 'サークルの話', '就活の情報交換'],
-    '趣味': ['ゲームしよう', 'カラオケ', '映画の話', '推し活'],
-    'その他': ['落とし物探して', 'ちょっと手伝って', '部室の掃除']
-}
-DAYS = ['月', '火', '水', '木', '金']
-# 簡易シラバスデータ
-COURSE_CONTENTS = ['Web工学', 'データベース', '人工知能', '線形代数', '心理学', '英語', '経済学入門', '物理学', '憲法', 'プログラミング基礎']
-
-def get_connection():
-    """設定に応じてSQLiteまたはPostgreSQLの接続を返す"""
-    if TARGET_DB_URL and TARGET_DB_URL.startswith("postgres"):
-        print(f"Connecting to PostgreSQL: {TARGET_DB_URL.split('@')[1]}") # ホスト名だけ表示
-        return psycopg2.connect(TARGET_DB_URL)
-    else:
-        print(f"Connecting to SQLite: {DB_PATH}")
-        return sqlite3.connect(DB_PATH)
-
-def execute_query(cur, query, params=()):
-    """SQLite(?)とPostgres(%s)のプレースホルダの違いを吸収して実行"""
-    if TARGET_DB_URL and TARGET_DB_URL.startswith("postgres"):
-        query = query.replace('?', '%s')
-    cur.execute(query, params)
-
-def init_db_structure(conn):
-    """DB構造を初期化する"""
-    cur = conn.cursor()
-    
-    # 既存テーブル削除 (Postgres用に CASCADE を追加)
-    tables = ["participants", "recruitments", "courses", "users", "custom_slots"]
-    for t in tables:
-        if TARGET_DB_URL: # Postgres
-             cur.execute(f"DROP TABLE IF EXISTS {t} CASCADE")
-        else: # SQLite
-             cur.execute(f"DROP TABLE IF EXISTS {t}")
-    
-    # テーブル作成 (ここは共通SQLでいける範囲で記述)
-    # Users
-    execute_query(cur, '''
-        CREATE TABLE users (
-            username TEXT PRIMARY KEY,
-            password TEXT NOT NULL,
-            grade INTEGER,
-            faculty TEXT,
-            department TEXT,
-            circles TEXT
-        )
-    ''')
-    # Courses
-    execute_query(cur, '''
-        CREATE TABLE courses(
-            username TEXT NOT NULL,
-            day      TEXT NOT NULL,
-            slot     INTEGER NOT NULL,
-            start_time TEXT,
-            end_time   TEXT,
-            content  TEXT,
-            PRIMARY KEY(username, day, slot)
-        )
-    ''')
-    # Recruitments
-    # Postgresでは AUTOINCREMENT は SERIAL or GENERATED BY DEFAULT AS IDENTITY だが
-    # 今回は SERIAL を使う書き方に統一するか、SQLite互換の書き方をする
-    # 簡易的に: id SERIAL PRIMARY KEY (Postgres), id INTEGER PRIMARY KEY AUTOINCREMENT (SQLite)
-    
-    if TARGET_DB_URL:
-        # PostgreSQL用 DDL
-        cur.execute('''
-            CREATE TABLE recruitments (
-                id SERIAL PRIMARY KEY,
-                creator_username TEXT NOT NULL,
-                title TEXT NOT NULL,
-                category TEXT,
-                max_participants INTEGER DEFAULT 2,
-                location TEXT,
-                date TEXT NOT NULL,
-                day TEXT NOT NULL,
-                start_slot INTEGER NOT NULL,
-                end_slot INTEGER NOT NULL,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-            )
-        ''')
-        cur.execute('''
-            CREATE TABLE participants (
-                id SERIAL PRIMARY KEY,
-                recruitment_id INTEGER NOT NULL,
-                applicant_username TEXT NOT NULL,
-                status TEXT DEFAULT 'pending',
-                applied_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (recruitment_id) REFERENCES recruitments(id)
-            )
-        ''')
-    else:
-        # SQLite用 DDL
-        cur.execute('''
-            CREATE TABLE recruitments (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                creator_username TEXT NOT NULL,
-                title TEXT NOT NULL,
-                category TEXT,
-                max_participants INTEGER DEFAULT 2,
-                location TEXT,
-                date TEXT NOT NULL,
-                day TEXT NOT NULL,
-                start_slot INTEGER NOT NULL,
-                end_slot INTEGER NOT NULL,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-            )
-        ''')
-        cur.execute('''
-            CREATE TABLE participants (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                recruitment_id INTEGER NOT NULL,
-                applicant_username TEXT NOT NULL,
-                status TEXT DEFAULT 'pending',
-                applied_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (recruitment_id) REFERENCES recruitments(id)
-            )
-        ''')
-
-    conn.commit()
-
-def create_demo_data():
-    print(f"--- デモデータ生成開始 (Seed: {RANDOM_SEED}) ---")
-    
-    conn = get_connection()
-    try:
-        init_db_structure(conn)
-        print("テーブル初期化完了")
-        cur = conn.cursor()
-
-        # ユーザー生成
-        print(f"ユーザー生成中 ({NUM_USERS}人)...")
-        usernames = []
-        common_hash = generate_password_hash("password")
-        
-        for i in range(NUM_USERS):
-            u = f"demo_user_{i+1}"
-            usernames.append(u)
-            grade = random.randint(1, 4)
-            faculty = random.choice(FACULTIES)
-            my_circles = random.sample(CIRCLES, random.randint(0, 3))
-            
-            execute_query(cur, 
-                "INSERT INTO users (username, password, grade, faculty, department, circles) VALUES (?,?,?,?,?,?)",
-                (u, common_hash, grade, faculty, "デモ学科", json.dumps(my_circles))
-            )
-
-        # 時間割生成
-        print("時間割生成中...")
-        for u in usernames:
-            num_courses = random.randint(8, 15)
-            slots = set()
-            while len(slots) < num_courses:
-                d = random.choice(DAYS)
-                s = random.randint(1, 8) 
-                slots.add((d, s))
-            
-            for d, s in slots:
-                content = random.choice(COURSE_CONTENTS)
-                execute_query(cur,
-                    "INSERT INTO courses (username, day, slot, start_time, end_time, content) VALUES (?,?,?,?,?,?)",
-                    (u, d, s, "00:00", "00:00", content)
-                )
-
-        # 募集生成
-        print(f"募集生成中 ({NUM_RECRUITMENTS}件)...")
-        recruitment_ids = []
-        today = datetime.now()
-        
-        insert_sql = """
-            INSERT INTO recruitments 
-            (creator_username, title, category, max_participants, location, date, day, start_slot, end_slot) 
-            VALUES (?,?,?,?,?,?,?,?,?)
-        """
-        # Postgresの場合 RETURNING id が使えるが、SQLite互換のため1件ずつINSERTしてカーソルからID取得
-        
-        for _ in range(NUM_RECRUITMENTS):
-            creator = random.choice(usernames)
-            category = random.choice(CATEGORIES)
-            title = random.choice(TITLES_BY_CAT.get(category, ['募集中']))
-            location = random.choice(LOCATIONS)
-            max_participants = random.randint(2, 6)
-            
-            date_offset = random.randint(-14, 14)
-            target_date = today + timedelta(days=date_offset)
-            date_str = target_date.strftime("%Y-%m-%d")
-            day_jp_map = ['月', '火', '水', '木', '金', '土', '日']
-            day = day_jp_map[target_date.weekday()]
-            
-            start_slot = random.randint(1, 8)
-            duration = random.choices([1, 2, 3], weights=[70, 20, 10])[0]
-            end_slot = min(start_slot + duration - 1, 8)
-            
-            if TARGET_DB_URL: # Postgres
-                cur.execute(insert_sql.replace('?', '%s') + " RETURNING id", (
-                    creator, title, category, max_participants, location, date_str, day, start_slot, end_slot
-                ))
-                rec_id = cur.fetchone()[0]
-                recruitment_ids.append(rec_id)
-            else: # SQLite
-                cur.execute(insert_sql, (
-                    creator, title, category, max_participants, location, date_str, day, start_slot, end_slot
-                ))
-                recruitment_ids.append(cur.lastrowid)
-
-        # 参加申請生成
-        print("参加データ生成中...")
-        for rec_id in recruitment_ids:
-            num_applicants = random.randint(0, 4)
-            if num_applicants == 0: continue
-            
-            applicants = random.sample(usernames, num_applicants)
-            for app_user in applicants:
-                status = random.choice(['pending', 'approved', 'rejected'])
-                execute_query(cur,
-                    "INSERT INTO participants (recruitment_id, applicant_username, status) VALUES (?,?,?)",
-                    (rec_id, app_user, status)
-                )
-
-        conn.commit()
-        print("完了: データベースにデータを生成しました。")
-
-    except Exception as e:
-        print(f"エラーが発生しました: {e}")
-        import traceback
-        traceback.print_exc()
-    finally:
-        conn.close()
-
-if __name__ == "__main__":
-    create_demo_data()
